[Added] Generate modal dialog for web
